backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
python-telegram-bot==20.0
openai==0.25.0
python-dotenv==0.21.0
black==22.12.0
pillow==9.3.0
ipython==8.7.0
discord.py==2.1.0
requests==2.28.1
pytest==7.2.0
aiogram==2.24
